feat(rendering): Adds color and texture updates from ECS to rendering layer
* Adds components that swap texture and color every 2 seconds
* Adds update materials data
* refactoring of some api for texture allocation, location definition
and update

+++ game/src/engine/ecs/time.rs
@@ -2,7 +2,7 @@ use bevy_ecs::prelude::*;
 
 #[derive(Resource, Default)]
 pub struct Time {
-    pub frames: f64,
+    pub frames: u64,
     pub time: f64,
     pub delta_time: f32,
     pub fixed_delta_time: f32
